Convert Utilities to Short, plus all references
Plugins left using long for now

+++ UCR.Core/Models/Plugin.cs
@@ -118,7 +118,7 @@ namespace HidWizards.UCR.Core.Models
 
         #region Plugin methods
 
-        protected void WriteOutput(int number, long value)
+        protected void WriteOutput(int number, short value)
         {
             Outputs[number].WriteOutput(value);
         }
