+++ UCR.Plugins/Remapper/AxisMerger.cs
@@ -34,9 +34,9 @@ namespace HidWizards.UCR.Plugins.Remapper
 
         public override void Update(params long[] values)
         {
-            var valueHigh = values[0];
-            var valueLow = values[1];
-            long valueOutput;
+            var valueHigh = (short)values[0];
+            var valueLow = (short)values[1];
+            short valueOutput;
 
             if (InvertHigh) valueHigh = Functions.Invert(valueHigh);
             if (InvertLow) valueLow = Functions.Invert(valueLow);
@@ -44,16 +44,16 @@ namespace HidWizards.UCR.Plugins.Remapper
             switch (Mode)
             {
                 case AxisMergerMode.Average:
-                    valueOutput = (valueHigh + valueLow) / 2L;
+                    valueOutput = (short) ((valueHigh + valueLow) / 2);
                     break;
                 case AxisMergerMode.Greatest:
                     valueOutput = Math.Abs(valueHigh) > Math.Abs(valueLow) ? valueHigh : valueLow;
                     break;
                 case AxisMergerMode.Sum:
-                    valueOutput = valueHigh + valueLow;
+                    valueOutput = (short) (valueHigh + valueLow);
                     break;
                 default:
-                    valueOutput = 0L;
+                    valueOutput = 0;
                     break;
             }
