+++ UCR.Plugins/Remapper/ButtonsToAxis.cs
@@ -16,7 +16,7 @@ namespace HidWizards.UCR.Plugins.Remapper
 
         public override void Update(params long[] values)
         {
-            long value;
+            short value;
 
             if (values[0] == 1 && values[1] == 1)
             {
