+++ UCR.Tests/UtilityTests/HelperTests/DeadZoneHelperTests.cs
@@ -17,7 +17,7 @@ namespace HidWizards.UCR.Tests.UtilityTests.HelperTests
         [TestCase(0, ExpectedResult = 0, TestName = "DeadZoneHelper (Init): 0 returns 0")]
         [TestCase(1, ExpectedResult = 1, TestName = "DeadZoneHelper (Init): 1 returns 1")]
         [TestCase(-1, ExpectedResult = -1, TestName = "DeadZoneHelper (Init): -1 returns -1")]
-        public long DeadZoneHelperInitTests(long inputValue)
+        public short DeadZoneHelperInitTests(short inputValue)
         {
             var helper = new DeadZoneHelper();
             return helper.ApplyRangeDeadZone(inputValue);
@@ -30,7 +30,7 @@ namespace HidWizards.UCR.Tests.UtilityTests.HelperTests
         [TestCase(-16384, 50, ExpectedResult = -1, TestName = "DeadZoneHelper (50): Half Negative deflection is outside DZ")]
         [TestCase(16383, 50, ExpectedResult = 0, TestName = "DeadZoneHelper (50): Below Half Positive deflection is inside DZ")]
         [TestCase(-16383, 50, ExpectedResult = 0, TestName = "DeadZoneHelper (50): Below Half Negative deflection is inside DZ")]
-        public long DeadZoneHelperValueTests(long inputValue, int percentage)
+        public short DeadZoneHelperValueTests(short inputValue, int percentage)
         {
             var helper = new DeadZoneHelper { Percentage = percentage };
             return helper.ApplyRangeDeadZone(inputValue);
